tb/requirements.txt
python-dotenv
pyyaml
pyhcl
pyfiglet
fuzzywuzzy
python-Levenshtein
requests
gitpython

boto3
PyCryptodome
prompt_toolkit